requirements.txt
qiskit==1.4.0
qiskit_aer==0.16.1
qiskit_dynamics==0.5.1
qiskit-ibm-runtime==0.36.1
qiskit[visualization]
python-dotenv
numpy
scipy
qiskit-experiments==0.10.0